nexpose_exploit_check.rb
A quick and dirty script and query to pull all known exploitable (or with malware kits) vulnerabilities within the last 30 days based on a multiple vulnerability name partial string matches and parses into a CSV file within the same directory and presorted by date descending. The string searches are based on the title of the vulnerability such as CVE’s, MS14-xxxx, php, etc. For performance reasons, please utilize the most specific queries possible. Modifying the script is trivial for no date restrictions or a large offset (however it may lead to very slow queries).